@JamesBoadi conversion: ComponentEdit.js to tsx
